Added posts & Added to Projects
Added the games to the projects and made a post for each game; they can be updated in the future, but for now this will work I think.

Note: It wouldn't let me run the website locally, so I don't know how this will actually look. Hoping for the best.